PROJECT -> edit and delete button

+++ apps/triplo/src/app/pages/project/project-detail/project-detail.component.ts
@@ -1,10 +1,13 @@
-import {Component, OnInit} from '@angular/core';
+import {Component, Inject, Injector, OnInit} from '@angular/core';
 import {ProjectHttpService} from "../../../services/projects/project-http.service";
 import {CommentInterface, ProjectInterface, TaskInterface} from "@triplo/models";
 import {ActivatedRoute, Router} from "@angular/router";
 import {CommentHttpService} from "../../../services/comments/comment-http.service";
-import {Observable} from "rxjs";
+import {PolymorpheusComponent} from '@tinkoff/ng-polymorpheus';
+import {Observable, switchMap} from "rxjs";
 import {TaskHttpService} from "../../../services/task/task-http.service";
+import {TuiAlertService, TuiNotification} from "@taiga-ui/core";
+import {ConfirmAlertComponent} from "../../../shared/alert/confirm/confirm-alert.component";
 
 @Component({
   selector: 'triplo-project-detail',
@@ -16,8 +19,12 @@ export class ProjectDetailComponent implements OnInit {
   comments$!: Observable<CommentInterface[]>
   id!: string
   $tasks: Observable<TaskInterface[]>;
+  notification: Observable<boolean>
 
   constructor(
+    @Inject(TuiAlertService)
+    private readonly alertService: TuiAlertService,
+    @Inject(Injector) private readonly injector: Injector,
     private route: ActivatedRoute,
     private router: Router,
     private projectService: ProjectHttpService,
@@ -35,14 +42,29 @@ export class ProjectDetailComponent implements OnInit {
     this.project$ = this.projectService.findProjectById(this.id)
     this.comments$ = this.commentService.getTopLevelComments(this.id)
     this.$tasks = this.taskService.getTopLevelTasks(this.id);
+
+
+    this.notification = this.alertService.open<boolean>(
+      new PolymorpheusComponent(ConfirmAlertComponent, this.injector),
+      {
+        label: `Are you sure you would like to delete this project?`,
+        status: TuiNotification.Error,
+        autoClose: false,
+      },
+    )
   }
 
   deleteProject() {
-    this.projectService.deleteProject(this.id).subscribe(
-      p => {
-        this.router.navigate(["/Projects"])
+    this.notification.subscribe(b => {
+      if (b) {
+        this.projectService.deleteProject(this.id).subscribe(
+          p => {
+            this.alertService.open('Deleted project', {label: "Success!"}).subscribe()
+            this.router.navigate(["/Projects"])
+          }
+        )
       }
-    )
+    })
   }
 
   back() {
